Add symbol.h
The preprocessor macro DANCY_SYMBOL(name) can be used for exporting
symbols. For example,

        void init(void)
        {

        }

        DANCY_SYMBOL(init);

The dy-link program handles the special "common" symbol so that "init"
is exported and there will be no memory overhead unless the compiler
settings force "common" to "bss". Even if that were the case, it would
still work but one "unsigned char" is put into the .bss section.