Remove lib objects
The acpica.at modules will be linked at run time. The whole lib module
and its exported symbols are usable. Currently, o64/acpica.at needs:

        isdigit         strcmp
        isprint         strcpy
        isspace         strlen
        isxdigit        strncmp
        memcmp          strncpy
        memcpy          tolower
        memset          toupper
        strcat

from the standard C library and all of these have been implemented. It
is very easy the get the list using the native linker:

        ./bin/dy-link ./o64/acpica.at --dump-ext

The actual kernel infrastructure is not ready but the list of "AcpiOs"
functions looks very reasonable. It is definitely doable to get ACPICA
running in the foreseeable future.