Introduce stack.c
For some time I have been envisioning an object model that takes some
ideas from C++ and its "this" pointer. This stack implementation is
the first experiment in refining the conventions of the Object Model.

        WARNING: this stack implementation is very inefficient!

It all starts from a base name, in this case the name is "stack" and
the following functions are always available:

        int stack_create(void **instance);
        int stack_delete(void *stack);

For the caller, there is an opaque data structure, and its name is:

        struct stack_instance;

All functions (or "methods") except the "stack_create" receive the
structure pointer and its name is the base name, "stack" in this case.
Callers never see the data structure definition, they simply use the
void pointer. To keep things simple, only "freestanding" headers can
be used and other library function prototypes are explicitly declared
at the beginning of the source file. For example:

        extern void free(void *ptr);
        extern void *malloc(size_t size);

Here the "extern" keyword is used for documenting that these function
declarations are outside of this compilation unit. The keyword usage
is redundant. Depending on the compiler, there might be some implicit
library function calls (e.g. memset) but those are not declared. The
function calls written explicitly must be declared here.

A full example of a program that uses this "stack" object:

        #include <stddef.h>
        #include <stdio.h>

        int stack_create(void **instance);
        int stack_delete(void *stack);

        int stack_push(void *stack, size_t size, const void *data);
        int stack_pop(void *stack, size_t size, void *data);

        int main(void)
        {
                char buffer[16];
                void *stack;

                if (stack_create(&stack))
                        return 1;

                stack_push(stack, 6, "world");
                stack_push(stack, 3, ", ");
                stack_push(stack, 6, "hello");

                while (!stack_pop(stack, 16, &buffer[0]))
                        printf("%s", &buffer[0]);
                printf("\n");

                stack_delete(stack);
                return 0;
        }

Please note that there is no header for this object. The user of this
object is free to have its own header structure but the object itself
is a "stand-alone" compilation unit. The unit only exports symbols
with the base name as a prefix, e.g. stack_create, stack_delete, etc.