Introduce error.h
It is time to introduce error codes. I think that this will make error
handling much more coherent. The error codes and messages are in one
place, and they should be very easy to keep synchronized.

   "X Macros are a technique for reliable maintenance of parallel
    lists, of code or data, whose corresponding items must appear
    in the same order. They are most useful where at least some of
    the lists cannot be composed by indexing, such as compile time."

                                                - Wikipedia

There is one special error, the "success", which is not an "error" but
means that everything is OK. The value of DE_SUCCESS must be zero.