Use constant size_t maximum values
I would definitely use SIZE_MAX but it is not always available. There
is one slight portability issue that I have not taken into account:
what if type size_t has smaller range than type unsigned int? At this
point, I am not aware of such implementations that can _also_ be used
for builing the Dancy tools. However, this is a "known issue" now and
I don't like it, but maybe this is getting too theoretical.