Introduce TrueType glyph rendering
This object introduces a simple interface for glyph rendering. It only
accepts "Dancy-compatible" TTF files (produced by the dy-ttf program),
although it should safely handle all input data.

        void ttf_demo(void)
        {
                extern unsigned char bitmap[32 * 32];

                extern void  *ttf_file_data;
                extern size_t ttf_file_size;

                unsigned int width;
                void *ttf;

                if (ttf_create(&ttf))
                        handle_error();

                if (ttf_open(ttf, ttf_file_size, ttf_file_data))
                        handle_error();

                if (ttf_set_bitmap(ttf, 32 * 32, &bitmap[0]))
                        handle_error();

                if (ttf_render(ttf, 0x0041u, &width))
                        handle_error();

                ttf_delete(ttf);
        }

The bitmap must be square. In the example above, the range of "width"
is from 0 to 32. Usually it is something like "half" of the square.