Introduce pre-kernel graphical user interface
The biggest problem with boot loader services is, obviously, the fact
that they are available only if the boot loader is still active. This
has been a major roadblock and I was not able to start configuring
hardware devices or installing interrupt handlers.

Now there is a module for outputting something on the screen so I can
start working on the code that is executed after the boot services has
been shut down and still have a way to show some messages to the user.

The windowing system is very simple but this is a good exercise before
starting to design and implement the real windowing system for the OS.
I'm not sure whether this "look & feel" is the final style guide, but
it could be. It can evolve when new features are implemented.