Add gui_move_window function
This was relatively easy to implement so I thought to try it. I will
need to study more windowing systems and decide what kind of strategy
would be best for my OS. At first, it intuitively feels that having a
compositing window manager that uses a bitmap buffer for each window
would be the simplest approach.