Implement anti-aliasing
The effect is applied to text rendering. The color palette needs to be
changed so that there will be more shades. In general, having smooth
graphics is better but there are some cases where it might not be the
preferred look, e.g. when not using the native resolution. This whole
thing is new to me so it might be that some adjustments are needed.