Fix anti-aliasing
The previous implementation had a bug that totally ignored the border
of the "em" box, i.e. the first and last pixel of each row, and also
the first and last row altogether. This modification slighly changes
the anti-aliasing behavior because now any pixel that was originally
turned on ("1") cannot be turned totally off. Only shades are allowed
for an originally set ("1") pixel of the monochrome bitmap.