Update TrueType glyph rendering
These fine adjustments complete the glyph renderer. I have studied the
TrueType virtual machine that would execute the font programs. Its own
language is very interesting, especially because I have used assembly
on the x86 and x86_64 platforms. Implementing the VM is not my goal at
this point but now I understand the problem it was meant to solve.

The ttf module is very small and trying to make it more complex could
weaken its greatest strength, the simplicity.