Remove B_FLAG_VALID_ENTRY
Although this is not a huge modification, I really like to get rid of
the "legacy" flag. The BIOS boot loader still sets the flag but now
the meaning has changed. Because the UEFI loader does not set it, we
can use this for detecting the boot loader type. The B_FLAG_UEFI flag
is not needed anymore.