Update memory.c
There are many difficulties if dynamically linked modules are too far
from each other in the memory space. I modified the pre-kernel memory
manager so that it does not allocate such memory areas. I assume that
all allocations should be relatively small anyway, so this should not
be a limitation. Please note that the manager is for the init module.