Fix overflows and wrong calling order
The minimum value for LONG_MAX is 0x7FFFFFFF and I accidentally used
the unsigned long int maximum value for lba range checks. This commit
fixes that real bug, but also it is now possible to have bigger file
system images if sizeof(long) is 8, for example.

I will think whether there should be preprocessing conditionals and
have 64-bit variants of fseek functions if those are available.

There was also a problem when calling fat_delete() because the image
file was closed before that. With FAT-32, the FS Info structure is
modified at that point.