Add fixed-dl option
I have read from many sources that some older BIOSes do not set the DL
register value. For smaller floppy disks, I modified the boot sector
instruction that originally wrote the DL value to BPB and now the DL
value is _read_ from the BPB. One theoretical drawback is that these
are not bootable from other drives than "0x00" unless the value in the
BPB is modified. I think this is a good compromise.

To make this even more complicated it could be possible that there are
some BIOS implementations that modify the BPB before control is given
to the boot code. In that case, the DL value in BPB would not be hard-
wired. I have not verified this but some writings suggest that. But I
think those systems are relatively modern and will not affect smaller
floppy disk sizes mentioned above.

I assume that 1.44 MB floppy and USB images are used on systems that
give the DL register value correctly. Especially for the latter, it
is very important to use the DL value BIOS gave because the value can
be almost anything depending on the boot configuration.