Write boot log files
This is a certain milestone because the OS writes something on a disk.
It does not work if using dancy.iso because that media is not writable
but floppies and USB images are. Boot loader services are used.

Log files are written in SYSTEM/ directory and are called:

        SYSTEM/LOG_INIT.TXT
        SYSTEM/LOG_UEFI.TXT

The latter is written if using the UEFI boot mechanism. I tested this
on an emulator and I did not find errors. However, I was a bit scared
to try this on my primary computer so I unplugged the real hard disk
drive before testing it. Everything seems to work.

I will concentrate on log file contents because there could be much
more interesting information. There is even more motivation to test
this system on real hardware because it is very easy to collect data
for further inspection.