Check memory allocation layout
This check should not be needed but in very unlikely scenarios it can
be possible that the second memory allocation is above the first one.
If I understood correctly, this might happen if the first allocation
used almost all the memory that was available and a tiny free memory
slot was available in the upper area of the 32-bit range. The OS would
not work anyway, so it is better to stop the boot process here.