Remove temporary code
This is a step forward. I will think about adding a background picture
because I already have a function for drawing it. As a bigger thing,
I have been thinking whether it makes sense to support VGA or not. It
is possible to keep it available in the pre-kernel environment but is
the complexity worth it in the actual OS environment? No, I think.

With good optimizations, the VGA performance could be quite good on
modern computers, but those computers most probably support VBE or GOP
linear framebuffer modes. There are checks for many CPU features that
make it even more unlikely that a computer that otherwise supported
the operating system would have VGA but no VBE/GOP support.