Modify b_exit functions
It is not possible to use boot loader services after calling "b_exit",
so I replace the syscall.asm .text section with int3 instructions. Not
exactly the full .text section because the new assembly routine itself
is not overwritten.

This modification is more relevant for the UEFI boot loader because an
accidental boot service call would have caused undefined behavior.