Add interrupt management
This simple implementation is for the pre-kernel environment after the
boot loader services are not usable anymore. At that time, the PICs
have been mapped by the boot loader, so IRQs start from vector 32, and
all of them have been disabled. If an I/O APIC is available, that will
be used but I have not developed code for it.

One thing worth mentioning is that I generate code for each interrupt
handler at run time. Those call the general handler, written in C. For
performance critical interrupts, it might be reasonable to replace the
common handler with a custom handler.