Terminate boot loader services
Calling the "b_exit" function can be considered to be a milestone. If
using the UEFI boot loader, this step is quite significant because a
relatively big part of firmware is terminated. Most of the available
memory in the system is owned by the operating system and the hardware
can be configured like operating systems usually do.

At this point, I load the GDT and set up interrupts.