Remove conditional statements
The new way to build ACPICA modules is to use the following command:

    make ACPICA_O32_AT=./o32/acpica.at ACPICA_O64_AT=./o64/acpica.at

I found make implementations that did not support "ifdefs". Now I have
tested this with "nmake", "make", and "bmake" and "pmake". It would be
possible to simplify the scripts if using some advanced make features.
However, I wanted to keep the scripts as simple as possible because of
compatibility. More importantly, if I wrote my own "dy-make", it would
need the basic features only.