Introduce build instructions
Please note the Disclaimer! It would be safe to do experiments on a
virtual computer, and this is true for both the build process and the
actual OS testing. Installing new software on a production computer is
always a risk. In this case I'm very confident that building the Dancy
Operating System itself is safe because it writes files only in the
source directory and does not even need root privileges. What is not
necessarily safe is installing the external programs (prerequisites).

Building the cross compiler is very time consuming, but fortunately it
is needed only once. All files are stored in the ./external/ directory.
It is possible to run "make" and "make clean" and the external tools
are not deleted. The "make distclean" fully restores the source tree.