Fix signed dword warning
The assembler printed a "signed dword value exceeds bounds" warning. I
intended to use the following instruction:

    AND r/m64, imm32        r/m64 AND imm32 (sign extended to 64-bits)

At the same time, I thought that it would be optimized to:

    AND r/m64, imm8         r/m64 AND imm8  (sign extended)

However, having 0xFFFFFFE0 as imm32 is correct but I agree that there
is a good reason for the assembler warning. The full disassembly for
the instruction "and rsp, 0xFFFFFFE0" is:

    48 83 E4 E0             and rsp, 0xFFFFFFFFFFFFFFE0

The assembler worked just like I thought. In general, I like to use
unsigned hexadecimals when handling bits, but in this case this does
not work very well. Sign extension is quite intrinsic part of 64-bit
instructions that have imm8 or imm32 operands. Based on this thought,
I came up with the obvious solution:

    and rsp, -32

This fixes the warning and is very obviously "AND r/m64, imm8" for
the assembler. The correspondence between the assemly source line and
the actual instruction is very strong.