Introduce paging
The memory map is identity mapped and it means that virtual addresses
equal physical addresses. As an important optimization, paging tables
can be created at run time (page fault exceptions are used).

The "pg_init" function maps all pages (except the first page) that are
below the end of "memory_map", so the code and data of the "init" is
mapped from the very beginnning. The overhead of those "static" tables
depends on the available free memory. If the computer does not have a
lot of memory, the address of "memory_map" is relatively low.

The pre-kernel environment can access any physical address except the
range from 0 to 4095. On 64-bit systems, it is possible to fill the
low memory with paging tables if trying to access too many different
addresses. It is not an undefined behavior but the memory allocation
would simply fail. The pre-kernel code is responsible for using only
physical addresses that contain RAM or memory-mapped devices.