Add gui_print_alert function
If there is an unrecovable error, e.g. "system panic", this function
can be used for printing a message. It is not guaranteed that it would
succeed in every possible scenario. If code and data memory areas were
corrupted, the system behavior would be undefined. It is intended that
the caller loads a "null interrupt descriptor table" so that nested
exceptions would cause a triple fault. It should be the safest option
and somewhat defined behavior.