Add idt_panic function
If there is an unhandled exception or interrupt, print a system panic
message. The pre-kernel environment doesn't try to recover the system.
The first 32 vectors are reserved ("exceptions") and vectors from 32
to 255 are "interrupts". However, the NMI is more like an interrupt.

The AMD64 Architecture Programmer's Manual (Volume 2) says:

   "Exceptions are considered synchronous events because they are
    a direct result of executing the interrupted instruction."

   "External interrupts are asynchronous events that occur
    independently of the interrupted instruction."

I try to be careful and use the correct terminology. The safest option
could be to use "vectors" from 0 to 255.