Add read and write functions
These simple functions do exactly what the caller says. Read or write
operations are not optimized away and unaligned memory addresses are
allowed. In C, calling a function is a sequence point and that is also
one of the reasons why these are implemented as separate functions.

These should not be used for performance-critical tasks. In some cases
it might be even desirable that read or write operations are a little
bit slower (e.g. memory mapped hardware). However, this kind of delay
is not very deterministic and proper delay functions must be used if
timing is critical.