Add rdmsr and wrmsr functions
These functions can read (rdmsr) and write (wrmsr) the model-specific
registers, which are like control registers. I have already used the
rdtsc instruction for reading the Time-Stamp Counter. The MSR for that
is IA32_TIME_STAMP_COUNTER (Intel's manual) or TSC (AMD's manual).

I try to use only MSRs that are "architectural" and supported by many
different CPU vendors. CPU Errata should be taken into account too,
but that is a huge topic in itself.

As a general thought, if some hardware were not compatible or behaved
erroneously, I would like to stop the OS initialization, assuming that
these cases can be reasonably detected. The first OS versions should
not go too far when it comes to workarounds or "hacks", although those
can be very important things in later phases.