Introduce spinlock functions
Spinlocks are simple synchronization primitives. If trying to acquire
an already held lock, the "spin_lock" function will keep checking the
lock value by using a normal "cmp" instruction. The "lock cmpxchg" is
not executed in a fast busy loop.

The other optimization is to use a "pause" instruction. The processor
manual (Intel) says that it avoids "the memory order violation in most
situations, which greatly improves processor performance" and it also
reduces the power consumption.

Although the concept is simple, using these functions correctly can be
tricky. Interrupt handlers, on the same CPU, must not acquire the lock
that is held by the code that gets interrupted. These kind of bugs are
hard to debug because the faulty conditions may not happen very often.