Introduce LAPIC and I/O APIC
This is the initial implementation of LAPIC and I/O APIC. There are a
lot of details and it might be that I need to modify something later,
after having more experience working with these. However, it is very
important to get started and hopefully I can initialize other CPUs in
the near future (Symmetric Multiprocessing).