Fix Boolean algebra
Unfortunately this was not the first time I made an error with boolean
logic. Often "OR" causes problems because "AND" is used more. In this
case, I had two "if" statements when I was writing the code but then I
combined them and did not pay much attention to it. This kind of logic
error could have been a major bug.