Initialize IRQ 0 clock source
Use HPET if that is available. If not, assume that PIT is available.
Current test code tries to enable RTC interrupts (IRQ 8) but that is
not going to work if using HPET. It is actually a good test because
those interrupts must not happen if HPET's "LegacyReplacement Option"
works like it should.

I have tested this only on one computer so I don't fully trust this
implementation at the moment. However, I'm very happy that in theory
the operating system is prepared for "legacy free" hardware.