Improve paging
Now it is possible to map a single page. Uncached page mappings will
use that feature. At the same time I found a major bug:

        page = (uint64_t)(addr & 0xFFE00000);

It definitely must be:

        page = (uint64_t)(addr & 0xFFFFFFFFFFE00000ull);

In other words, identity mappings for addresses above 0xFFFFFFFF did
not work as intended. The behavior was defined but wrong. The physical
memory area from 0 to 0xFFFFFFFF was simply mapped again, and that was
not "identity mapping" at all.