Introduce delay function
My implementation uses Time-Stamp Counter (TSC). The biggest challenge
is to calibrate the delay function, and it might be that its frequency
changes. It is surprisingly hard to implement a good delay function. I
even thought whether I should have a simple busy loop (without TSC).

The delay function uses "uint32_t nanoseconds" and this was on purpose
because the function should not be used for longer "sleeps". Although
this is the pre-kernel environment, I will probably use this interface
from now on. For many cases it might be too accurate.