Introduce Symmetric Multiprocessing
This is a milestone and I'm very happy that I have this available now.
If I had continued coding without this for any longer, it would have
been a nightmare to retrofit it later. Now I can take SMP into account
all the time. Most of the current init code is not SMP-compatible but
I will figure out what needs to be updated. Perhaps not much because
the pre-kernel environment (init) will be replaced by the kernel.

It took a few late evenings to write this feature, although the amount
of code is not very huge. This time I was a bit surprised that almost
the first test somewhat worked and APs woke up.

A known optimization is to shorten the delays I used when sending the
Interprocessor Interrupts. I'll keep it like this for now and optimize
later if needed. Correctness and reliability are the important things
and I try to test this on real hardware as much as I can.