Use wbinvd instructions
When I added read and write functions, I wrote in the commit message
that they should not be used for performance-critical tasks. Today I
make the write functions even slower by adding wbinvd instructions.
Now, calling a write function is a sequence point in C, and it also
executes a serializing instruction (wbinvd).

The wbinvd instruction does more than serializing, but the end result
is that the caller can write bytes, words, or dwords so that caching
and out-of-order execution effects have been practically removed.