Rethink error handling
It seems that I have not used the "standard" error codes in the init
module so I removed them.

Today I have been doing refactoring for header files. The plan is to
start designing what comes after the init module. The "dancy.h" header
is going to be included and I try to keep the namespace under control.
For example, the "boot.h" header has declarations for b_output_string
and similar functions that are not used after the init module.