Use b_print function pointer
When the pre-kernel GUI is initialized, boot loader services are not
used for printing. Function pointers are very good, although they add
some overhead, but in this case the performance is not very relevant.

Actually, the added overhead is totally irrelevant when comparing the
time spent on actual printing. Changes in object relocations may have
a bigger impact on total performance, but still insignificant. It is
good to think aspects like these once in a while. For a critical task
it could make a measurable difference.