Improve Symmetric Multiprocessing
The smp_ap_count variable can give up-to-date information immediately
after the smp_init function returns. Before this commit, it was almost
certain that APs run the smp_ap_entry in a defined order. However, it
was not a good assumption.

Now all APs should wait for the spinlock just before it is unlocked. I
like the idea that "randomness of SMP" is present here because it can
reveal possible bugs in later phases.