Revise ld.c
I increased the name buffer size. Modern C standards allow 31 initial
characters in an external identifier. The new size is bigger than this
and the magic size of 36 is for alignment. The last char must be NUL.

Now it is possible to access the symbol table directly. Most probably
I will copy the whole table after the kernel has been linked.