Enable AP interrupts
Application processors use the same interrupt handling mechanism. The
bootstrap processor handles "panic() messages" also for APs. This kind
of simple arrangement would not be very good for a real kernel, but it
is suitable for the pre-kernel environment.

Now, if an AP is halted, it can be woken up by sending an IPI and the
vector can be from 32 to 255. It would have been possible to use NMIs
for the same purpose if interrupts need to be disabled.