Extend spinlock functions
The "spin_lock" function does not disable interrupts. I am introducing
a function called "spin_enter" which saves the interrupt flag and then
disables local interrupts. Then it calls the "spin_lock" function. The
other function is called "spin_leave" which restores the interrupts so
that the original state is preserved.

An example of using these new functions:

        int example_lock;

        void example_function(void)
        {
                void *example_lock_local = &example_lock;

                /*
                 * Acquire the lock.
                 */
                spin_enter(&example_lock_local);

                /*
                 * Critical section with interrupts disabled.
                 */
                example_process_data();

                /*
                 * Release the lock.
                 */
                spin_leave(&example_lock_local);
        }

The lock variable type is int and on supported platforms it has 4-byte
alignment. It is possible to use the low bits for saving information.