Introduce simple threads
Having threads in the pre-kernel environment would not be necessary,
but I wanted to use this opportunity to gain experience. Later I need
to implement these in the actual kernel. Unfortunately this pre-kernel
was not designed to be thread-safe (not to mention SMP safeness), so I
can't do complicated things with multiple threads right now.

I noticed this same thing when starting up application processors. It
is not reasonably possible to retrofit multiprocessing in an existing
environment. Of course, it can be done but the end result is not very
elegant even if bugs were not a big challenge in themselves.

Multiprocessing must be taken into account from the very beginning and
I will do that when developing the kernel. It is very good that I have
a pre-kernel environment with these features somewhat available so the
mindset will be already established before starting the kernel.