Improve Symmetric Multiprocessing
I did not notice that if all bits are set in the destination APIC ID,
the interrupt is broadcasted and accepted by all APICs. To make this
more interesting, older processors may support 4 bits only. As far as
I know, 32-bit systems with more than 15 different CPUs are rare. I
solved the problem by simply limiting the allowed value if compiling
the 32-bit version of the operating system.

I changed some comments because the second SIPI will not sent if the
first SIPI already started the AP. In addition, it is a bit hard to
guarantee that everything works in a defined way if the AP suddenly
starts after the "smp_init" function thought that it failed. For this
reason, about two seconds will be spent on waiting the reply. It is
a lot of time, but this hopefully guarantee that the AP really is not
working and will not wake up later.