Add spin_trylock functions
It is a bit misleading to have a word "spin" here in this context. The
function does not actually spin but the value is returned immediately.
If the return value is zero, the lock was not acquired. Example:

        if (spin_trylock(&lock)) {
                /*
                 * The lock was acquired.
                 */
                do_something();
        } else {
                /*
                 * The lock was not acquired.
                 */
                do_something_else();
        }

This function should have been implemented from the very beginning. It
is very useful to give the caller an option to do something else. The
spin_lock function gives control away from the caller.