Introduce mutex functions
It was not necessary to have threads in the pre-kernel environment. In
the same way, mutex functions would not be needed. However, this was a
very good lesson and I try to gain experience as much as possible when
still working on this "init" module.

The interfaces are defined in the C11 standard. Because I haven't used
these in real coding I can't review them at this point. The constants
as return values (e.g. thrd_success) can be something that I'm not too
convinced of. It would be simpler if "success" was always zero. In my
implementation it happens to be zero, but it would not be a good idea
to write code based on that implementation detail.