Improve mutex functions
Only the same thread of execution can lock the mutex recursively, if
the mutex type is mtx_recursive. I'm not sure whether it is a good
idea to have recursive mutex objects in general. It may help in some
cases, but I think that it most probably hides bugs.

It could make sense to use the panic() function if locking rules are
not followed. Systems programming is critical, and having undefined
behavior could be worse than defined system crashes.