Delay NMI handling
A non-maskable interrupt can happen at any time. The handler just sets
a flag if an NMI occurs. The flag will be checked when handling normal
interrupts (vectors from 32 to 255). Most likely it will be the IRQ 0.

The pre-kernel environment does not try to handle NMIs, but these are
unrecoverable errors. The idt_panic() function is used.