Change idt_handler interface
The err_code parameter is not needed. It is possible to read the error
code from the stack. The stack parameter has a const qualifier because
this "init" module does not support system calls. Otherwise, it could
be used for changing the values that will be popped when returning. It
is a common convention to use software-generated interrupts for system
calls (especially on 32-bit systems), and then return values would be
very important.

It is a very serious bug if an asynchronous interrupt changes register
values. Having undefined behavior is a problem in itself but debugging
those kind of "random" register value corruptions might be very hard.