Use mutual exclusion functions
In the same way as with the memory functions, the "GUI services" need
to be thread-safe. It is possible to use these services before calling
the b_exit() function, and that is why function pointers are utilized.

Code refactoring was needed. Also, I decreased the "gui_print" buffer
size because current stack sizes for threads resulted in overflows. I
have to check and verify how much stack space is used, and have proper
safety margins in other places too.