Delete init_ap_error variable
It was not a good idea to have this kind of "hack". A proper solution
is to improve the panic() function so that APs can call it. Actually,
they can call it right now but it does not print the error message.

As other changes, now nested exceptions are allowed and the buffer in
the idt_panic function is local again.