Delete delayed NMI handling
An application processor can send a non-maskable interrupt to the BSP.
There are many problems if the BSP keeps going for a while. Today I've
made a lot of changes in file "idt.c", and it looks much better now.

Because NMIs can happen anytime, the gui_print_alert function is more
critical. It should print the alert message but some kind of graphical
glitches may be in the background if an NMI happens at a bad time. All
these are unrecoverable errors, so that drawback should be acceptable.