Modify thrd_yield function
Although application processors do not support threads at the moment,
I definitely want that they can safely call the thrd_yield() function.
It could make sense to have the general part written in C, but it was
quite easy to implement the safety test in assembly. The CPU-specific
part would be needed for pushing/popping the non-volatile registers.