Improve Symmetric Multiprocessing
The second INIT (De-assert) is not needed on computers that otherwise
meet the requirements to run this smp_init function. One of reasons I
definitely want this removed is because the Intel manual says:

101 (INIT Level De-assert)
        ...
        This IPI is sent to all processors, regardless of the
        value in the destination field; however, software should
        specify the "all including self" shorthand.

If I understood correctly, that behavior is for older processors. The
same manual mentions that the "Level flag" is always issued as a 1 if
running this on modern CPUs. This probably means that the same normal
INIT would be sent twice. It does not sound good at all, especially
because the delay between the second INIT and the SIPI was short.