Improve Symmetric Multiprocessing
The stack size for application processors has been increased. I made
many changes that try to cope with the uncertainty in the cases where
APs do not start up in an expected way. It is not very easy to find
out are they running or not. It is possible that there is a bug that
prevents them from setting up the status value although they started
up correctly.

This operating system is meant to run on CPUs that do not require the
second Startup IPI when starting up the APs. By saying this I try to
justify the slowness of the code path that sends the second SIPI.

However, the good thing is that all this complexity is in one place.
The other systems can assume that APs are in a stable state after the
smp_init() function has finished.