Improve Symmetric Multiprocessing
In the previous commit message I worried about the case where the APs
were started up but the status value was not changed due to unknown
bugs. Now the INIT (Assert) is sent to those APs, and then a warning
message is printed. Another message, that is not an error, is printed
when two Startup IPIs were needed for starting up some or all APs. If
this latter message is printed too often, I need to optimize the code.
Delays have been set so that sending the second Startup IPI is not a
normal and expected code path.