Refine system directory
I did code refactoring some time ago and the system directory became a
permanent part of the Dancy source tree. This was not a good idea and
I totally forgot the original vision I had. The system directory must
contain only the files that are copied on the bootable media and it is
simpler to copy the whole directory instead of individual files.

For example, the files on a bootable FAT16 formatted media could be:

    EFI
      BOOT
        BOOTX64.EFI

    SYSTEM
      CONFIG.AT
      DB_000.AT
      DB_001.AT
        ...
      IN_IA16.AT
      IN_IA32.AT
      IN_X64.AT

    LOADER.512
    LOADER.AT

The above hierarchy that is on the bootable media follow exactly the
same structure as the Dancy source tree itself, after everything has
been built. The required files for booting are a subset of all files
of the (built) source tree, but the full tree can be stored also.