Add pci.c
The enhanced configuration mechanism is used if that is available. It
would be interesting to try this on real hardware with more than one
PCI Segment Group. These functions do not currently modify the state
of the devices but I would like to disable "USB Legacy Support". The
real PS/2 Controller (if available) can not be initialized if there
is some kind of emulation mode enabled.

Support for USB is definitely the goal. It might take some time until
I have enough knowledge to implement that. Things are not necessarily
hard but they simply require work.