Develop pci.c
I added new structure members and functions. The b_log format has been
changed. It is very interesting to run this on different computers and
read the log files.

The first parameter of the pci_read function has a const qualifier but
the write function has not. Although the structure is not modified, it
looked a bit confusing to have the const qualifier. It looks familiar
to see it associated with "read" but not with "write".