Improve interrupt handling
Today I tried my operating system on different computers. One computer
did not trigger timer interrupts and after some debugging I found out
that it used I/O APIC but the ACPI tables did not have the "Interrupt
Source Override Structure" for "ISA IRQ 0 to Input 2". However, it had
that mapping because IRQ 2 was triggered instead of IRQ 0.

Also, the HPET specification says that the "Legacy Replacement" will
force the mapping of "IRQ 0 to IRQ 2". Other computer on which I have
tried this have had the correct Override Structure in the MADT table.