Use uncached configuration space
If the Enhanced Configuration Access Mechanism (ECAM) is used, map the
memory areas as uncached before doing any writes. The initialization
function itself does not write into the configuration space. It would
be more elegant to do this earlier but the pci_init_early function is
still compatible with the BIOS boot loader. For the UEFI boot loader,
all configuration space accesses are done after the ExitBootServices()
function. In that case the pci_init_early function just allocates the
memory but does not do the PCI enumeration.