Add usb_init_ehci function
The "Pre-OS to OS Handoff" routine is interesting. The specification
contains quite detailed description. Hopefully there are not too many
BIOS implementations that fail this handoff routine. For operating
system developers, this "System Management Mode" is a bit difficult
topic. It gives valuable features but, on the other hand, it is hard
to know exactly what is happening at a given time. And firmware bugs.